Move getStartOfYear.js to luxon dir
