+++ app/scripts/actions/就業年報画面に今年ボタンを追加.js
@@ -1,4 +1,4 @@
-import getStartOfYear from '../modules/getStartOfYear'
+import getStartOfYear from '../modules/luxon/getStartOfYear'
 import getNow from '../modules/luxon/getNow'
 
 export default function () {
